+++ e2e-tests/runner/examples/connectors/WalletConnectConnector/WalletConnectConnector.test.ts
@@ -47,10 +47,12 @@ test.describe('WalletConnectConnector', () => {
   // First-time connection requires a message signature (to prove ownership of the wallet)
   const connect: ConnectorFunctions['connect'] = async (page) => {
     await commonConnect(page);
+    await page.getByText('Sign', { exact: true }).click();
     await metamask.confirmSignature();
+    await page.getByText('Continue to application').click();
   };
 
-  // From here on, we'll skip the signature step
+  // From here on, we'll skip the signature step and predicate address warning disclaimer
   const secondConnect: ConnectorFunctions['connect'] = commonConnect;
 
   const approveTransfer: ConnectorFunctions['approveTransfer'] = async () => {